wallet: path handling.
