+++ lib/wallet/txdb.js
@@ -403,7 +403,7 @@ TXDB.prototype.commit = co(function* commit() {
  */
 
 TXDB.prototype.getInfo = function getInfo(tx) {
-  return this.walletdb.getPathInfo(this.wallet, tx);
+  return this.wallet.getPathInfo(tx);
 };
 
 /**
