+++ lib/wallet/wallet.js
@@ -28,6 +28,7 @@ var MasterKey = require('./masterkey');
 var Input = require('../primitives/input');
 var Output = require('../primitives/output');
 var LRU = require('../utils/lru');
+var PathInfo = require('./pathinfo');
 
 /**
  * BIP44 Wallet
@@ -1822,6 +1823,37 @@ Wallet.prototype.getTX = function getTX(hash) {
 Wallet.prototype.addTX = function addTX(tx) {
   return this.db.addTX(tx);
 };
+
+Wallet.prototype.add = co(function* add(tx) {
+  var info = yield this.getPathInfo(tx);
+  yield this.txdb.add(tx, info);
+  yield this.handleTX(info);
+});
+
+Wallet.prototype.unconfirm = co(function* unconfirm(hash) {
+  return yield this.txdb.unconfirm(hash);
+});
+
+/**
+ * Map a transactions' addresses to wallet IDs.
+ * @param {TX} tx
+ * @returns {Promise} - Returns {@link PathInfo}.
+ */
+
+Wallet.prototype.getPathInfo = co(function* getPathInfo(tx) {
+  var hashes = tx.getHashes('hex');
+  var paths = [];
+  var i, hash, path;
+
+  for (i = 0; i < hashes.length; i++) {
+    hash = hashes[i];
+    path = yield this.getPath(hash);
+    if (path)
+      paths.push(path);
+  }
+
+  return new PathInfo(this, tx, paths);
+});
 
 /**
  * Get all transactions in transaction history (accesses db).
